# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Threat Infrastructure Mapper (TIM) API",
    "description": "Local-first threat intelligence and infrastructure mapping. Submit one IOC and TIM collects, fingerprints, enriches, pivots and correlates the infrastructure that belongs to the same operation.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/investigate": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Start an investigation from a single IOC (domain, URL, IP or certificate hash)",
        "operationId": "investigate_investigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/bulk-investigate": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Start investigations for up to 500 IOCs",
        "operationId": "bulk_investigate_bulk_investigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkInvestigationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkInvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/investigations": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "List investigations",
        "operationId": "list_investigations_investigations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on IOC",
              "title": "Q"
            },
            "description": "Substring match on IOC"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvestigationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "bulk_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bulk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InvestigationListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Investigation detail and stage progress",
        "operationId": "get_investigation_investigation__inv_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Investigations"
        ],
        "summary": "Delete an investigation and its artifacts",
        "operationId": "delete_investigation_investigation__inv_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/provider-runs": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Every provider call made by the investigation",
        "operationId": "provider_runs_investigation__inv_id__provider_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Provider Runs Investigation  Inv Id  Provider Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/assets": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Assets found by an investigation",
        "operationId": "investigation_assets_investigation__inv_id__assets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AssetListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/artifacts": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Raw collected artifacts",
        "operationId": "investigation_artifacts_investigation__inv_id__artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactOut"
                  },
                  "title": "Response Investigation Artifacts Investigation  Inv Id  Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/cancel": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Cancel a queued/running investigation",
        "operationId": "cancel_investigation_investigation__inv_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/rerun": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Re-run an investigation with the same IOC and options",
        "operationId": "rerun_investigation_investigation__inv_id__rerun_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/investigation/{inv_id}/notes": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Add an analyst note",
        "operationId": "add_note_investigation__inv_id__notes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "inv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inv Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationNote"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkInvestigationCreate": {
        "properties": {
          "iocs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Iocs"
          },
          "options": {
            "$ref": "#/components/schemas/InvestigationOptions"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          }
        },
        "type": "object",
        "required": [
          "iocs"
        ],
        "title": "BulkInvestigationCreate"
      },
      "BulkInvestigationOut": {
        "properties": {
          "bulk_id": {
            "type": "string",
            "title": "Bulk Id"
          },
          "investigation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Investigation Ids"
          },
          "rejected": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Rejected"
          }
        },
        "type": "object",
        "required": [
          "bulk_id",
          "investigation_ids",
          "rejected"
        ],
        "title": "BulkInvestigationOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IOCType": {
        "type": "string",
        "enum": [
          "domain",
          "url",
          "ip",
          "certificate"
        ],
        "title": "IOCType"
      },
      "InvestigationCreate": {
        "properties": {
          "ioc": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Ioc",
            "examples": [
              "example.com"
            ]
          },
          "options": {
            "$ref": "#/components/schemas/InvestigationOptions"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          }
        },
        "type": "object",
        "required": [
          "ioc"
        ],
        "title": "InvestigationCreate"
      },
...
      "InvestigationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ioc": {
            "type": "string",
            "title": "Ioc"
          },
          "ioc_type": {
            "$ref": "#/components/schemas/IOCType"
          },
          "normalized": {
            "type": "string",
            "title": "Normalized"
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          },
          "options": {
            "$ref": "#/components/schemas/InvestigationOptions"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StageState"
            },
            "type": "object",
            "title": "Stages"
          },
          "summary": {
            "$ref": "#/components/schemas/InvestigationSummary"
          },
          "root_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Asset Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "bulk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bulk Id"
          },
          "cluster_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cluster Ids"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "notes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ioc",
          "ioc_type",
          "normalized",
          "status",
          "options",
          "stages",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "InvestigationOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token from POST /auth/login",
        "scheme": "bearer"
      }
    }
  },
...
}